refactor: edit Search page
